backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["m-shao <dev.minglunshao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-texttospeech = "^2.17.2"
pygame = "^2.6.0"
levenshtein = "^0.25.1"
scikit-learn = "^1.5.2"
speechrecognition = "^3.10.4"
pyaudio = "^0.2.14"
sounddevice = "^0.5.0"
pydub = "^0.25.1"
wheel = "^0.44.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
